When removing a body, only test attached bodies for grabs
When removing a body from the environment, we check whether any bodies
were still grabbing it and, if so, ungrab it before deleting it.
However, we scan the entire set of bodies in the environment to check
if they are grabbing, which is a lot of overhead for environments with a
lot of tracked bodies.
Since all grab relations seem to also count as attached body relations,
limit the search for grab partners to only the set of bodies that are
directly attached to the body to be removed.